requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tavily-python==0.7.10
langchain==0.3.27
langchain-core==0.3.72
langchain-anthropic>=0.1.23
langgraph>=0.2.6
langgraph-checkpoint>=2.1.0
pydantic==2.11.7
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.28.1
typing-extensions>=4.0.0